pyproject.toml
[project]
name = "training-tracker"
version = "0.1.0"
description = ""
authors = [
    {name = "Anastasiya Avseikova",email = "muursea@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "django-filter (>=25.1,<26.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "signals (>=0.0.2,<0.0.3)",

]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
